ui and logic is working
